+++ src/constants.ts
@@ -49,6 +49,13 @@ export const ENABLE_PERSON_PROCESSING = '$epp'
 export const TOOLBAR_ID = '__POSTHOG_TOOLBAR__'
 export const TOOLBAR_CONTAINER_CLASS = 'toolbar-global-fade-container'
 
+/**
+ * PREVIEW - MAY CHANGE WITHOUT WARNING - DO NOT USE IN PRODUCTION
+ * Sentinel value for distinct id, device id, session id. Signals that the server should generate the value
+ * */
+export const COOKIELESS_SENTINEL_VALUE = '$posthog_cklsh'
+export const COOKIELESS_MODE_FLAG_PROPERTY = '$cklsh_mode'
+
 export const WEB_EXPERIMENTS = '$web_experiments'
 
 // These are properties that are reserved and will not be automatically included in events