+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.201.0 - 2024-12-16
+
+- fix: rotate session id proactively (#1512)
+
+## 1.200.2 - 2024-12-16
+
+- feat(web-analytics): Add flag to send server hash instead of distinct id (#1490)
+
 ## 1.200.1 - 2024-12-13
 
 - fix: Make legacy property sturdier (#1602)